revert the changes that make Downloader an interface and rename Downloader to AbstractDownloader. Move application status code getters into DownloaderProperties
