+++ hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/downloader/DownloaderProperties.java
@@ -21,10 +21,10 @@ package com.hedera.mirror.importer.downloader;
  */
 
 import java.nio.file.Path;
-import java.time.Duration;
 import javax.annotation.PostConstruct;
 
 import com.hedera.mirror.importer.MirrorProperties;
+import com.hedera.mirror.importer.domain.ApplicationStatusCode;
 import com.hedera.mirror.importer.domain.StreamType;
 import com.hedera.mirror.importer.util.Utility;
 
@@ -67,6 +67,10 @@ public interface DownloaderProperties {
     boolean isKeepSignatures();
 
     void setKeepSignatures(boolean keepSignatures);
+
+    ApplicationStatusCode getLastValidDownloadedFileKey();
+
+    ApplicationStatusCode getLastValidDownloadedFileHashKey();
 
     @PostConstruct
     default void init() {
